+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -152,11 +152,8 @@ class NetCDFWrite(IOWrite):
     #--- End: def
     
     def _write_attributes(self, parent, ncvar, extra={}, omit=()):
-        '''
-:Examples 1:
+        '''TODO
 
->>> w._write_attributes(x, 'lat')
-    
 :Parameters:
 
     parent:
@@ -171,8 +168,6 @@ class NetCDFWrite(IOWrite):
 
     `None`
 
-:Examples 2:
-    
         '''
         if parent is None:
             netcdf_attrs = {}
@@ -186,6 +181,17 @@ class NetCDFWrite(IOWrite):
     
         if not netcdf_attrs:
             return
+
+        # Make sure that _FillValue and missing data have the same
+        # data type as the data
+        for attr in ('_FillValue', 'missing_value'):
+            if attr not in netcdf_attrs:
+                continue
+
+            data = self.implementation.get_data(parent, None)
+            if data is not None:
+                netcdf_attrs[attr] = numpy.array(netcdf_attrs[attr], dtype=data.dtype)
+        #--- End: for
         
         self.write_vars['nc'][ncvar].setncatts(netcdf_attrs)
     #--- End: def